Clean out e2e-tools
